added cache to paper object
I added a cache to the paper object, so that searching for the same paper twice doesn't result in an extra query, making it use the limited API calls less. This is mostly useful for tests. The current implementation isn't probably somewhat suboptimal (I mean I hold all the searched papers in the paper object, which means it duplicates the library info... So the next step is to clean this up)

Also reconfigured tests so they don't anger the linter anymore