beginnings of a gui
This is the foundation of a GUI, with the basic shape I want. It will need to be refactored to make things talk to each other. And there are no tests for this written now. I waited on those to see what I actually wanted to do and what is possible with qt