added arXiv identifier to database
This will later be used to show the arXiv id for yet-unpublished papers

The ADSWrapper class now parses the paper identifiers to get this info, and it is stored in the database. As some papers may not be on the arXiv, we test for that possibility with a newly added paper.